+++ src/main/java/chien/demo/shopdemo/service/impl/CartServiceImpl.java
@@ -1,53 +1,64 @@
 package chien.demo.shopdemo.service.impl;
 
-import chien.demo.shopdemo.dto.CartDTO;
+import chien.demo.shopdemo.dto.CartDto;
 import chien.demo.shopdemo.mapper.CartMapper;
 import chien.demo.shopdemo.mapper.CustomerMapper;
 import chien.demo.shopdemo.model.Cart;
 import chien.demo.shopdemo.repository.CartRepository;
 import chien.demo.shopdemo.service.CartService;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-
+/** The type Cart service. */
 @Service
 @Transactional(rollbackFor = Throwable.class)
 public class CartServiceImpl implements CartService {
 
-    @Autowired
-    private CartRepository cartRepository;
+  @Autowired private CartRepository cartRepository;
 
-    @Override
-    public List<CartDTO> findAll() {
-        return cartRepository.findAll().stream().map(cart ->
-                CartMapper.getInstance().toDTO(cart)).collect(Collectors.toList());
-    }
+  @Override
+  public List<CartDto> findAll() {
+    return cartRepository.findAll().stream()
+        .map(cart -> CartMapper.getInstance().toDto(cart))
+        .collect(Collectors.toList());
+  }
 
-    @Override
-    public CartDTO create(CartDTO dto) {
-        Cart cart = CartMapper.getInstance().toEntity(dto);
-        return CartMapper.getInstance().toDTO(cart);
-    }
+  @Override
+  public CartDto create(CartDto dto) {
+    Cart cart = CartMapper.getInstance().toEntity(dto);
+    return CartMapper.getInstance().toDto(cart);
+  }
 
-    @Override
-    public CartDTO update(int id, CartDTO dto) {
-        Cart cart = cartRepository.findById(id).get();
-        cart.setCustomer(CustomerMapper.getInstance().toEntity(dto.getCustomer()));
-        return CartMapper.getInstance().toDTO(cart);
-    }
+  @Override
+  public CartDto update(int id, CartDto dto) {
+    Optional<Cart> result = cartRepository.findById(id);
+    Cart cart = result.orElseGet(Cart::new);
+    cart.setId(id);
+    cart.setCustomer(CustomerMapper.getInstance().toEntity(dto.getCustomer()));
+    return CartMapper.getInstance().toDto(cart);
+  }
 
-    @Override
-    public void delete(int id) {
-        Cart cart = cartRepository.findById(id).get();
-        cartRepository.delete(cart);
+  @Override
+  public void delete(int id) {
+    Optional<Cart> result = cartRepository.findById(id);
+    Cart cart;
+    if (result.isPresent()) {
+      cart = result.get();
+      cartRepository.delete(cart);
     }
+  }
 
-    @Override
-    public CartDTO findById(int id) {
-        return CartMapper.getInstance().toDTO(cartRepository.findById(id).get());
+  @Override
+  public CartDto findById(int id) {
+    Optional<Cart> result = cartRepository.findById(id);
+    if (result.isPresent()) {
+      return CartMapper.getInstance().toDto(result.get());
+    } else {
+      return null;
     }
+  }
 }
